+++ src/diff.c
@@ -2844,7 +2844,7 @@ ex_diffgetput(exarg_T *eap)
 	    // Adjust marks.  This will change the following entries!
 	    if (added != 0)
 	    {
-		mark_adjust(lnum, lnum + count - 1, (long)MAXLNUM, (long)added);
+		mark_adjust(lnum, lnum + count - 1, MAXLNUM, (long)added);
 		if (curwin->w_cursor.lnum >= lnum)
 		{
 		    // Adjust the cursor position if it's in/after the changed
