+++ src/help.c
@@ -561,7 +561,7 @@ find_help_tags(
     flags = TAG_HELP | TAG_REGEXP | TAG_NAMES | TAG_VERBOSE | TAG_NO_TAGFUNC;
     if (keep_lang)
 	flags |= TAG_KEEP_LANG;
-    if (find_tags(IObuff, num_matches, matches, flags, (int)MAXCOL, NULL) == OK
+    if (find_tags(IObuff, num_matches, matches, flags, MAXCOL, NULL) == OK
 	    && *num_matches > 0)
     {
 	// Sort the matches found on the heuristic number that is after the
